Issue #1970 Moving integration test code to a separate go package
