+++ Makefile
@@ -64,7 +64,7 @@ DOCS_UID ?= $(shell docker run -tiv $(LOCAL_DOCS_DIR):$(CONTAINER_DOCS_DIR) $(DO
 DOC_VARIABLES = -e OPENSHIFT_VERSION=$(OPENSHIFT_VERSION) -e MINISHIFT_VERSION=$(MINISHIFT_VERSION) -e CENTOS_ISO_VERSION=$(CENTOS_ISO_VERSION) -e MINIKUBE_ISO_VERSION=$(MINIKUBE_ISO_VERSION)
 
 # MISC
-START_COMMIT_MESSAGE_VALIDATION = 204ce1930e8a86d0c5b5eabd1d8b8dbf8d9edac3
+START_COMMIT_MESSAGE_VALIDATION = 83af4780db01d9a4607925af83297e32b38a1603
 
 # Check that given variables are set and all have non-empty values,
 # die with an error otherwise.
